third_party/blink/web_tests/platform/generic/fast/forms/number/number-spinbutton-capturing-expected.txt

Test for a capturing bug of spin buttons in a type=numnber input.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS anotherInput.value is "2"
PASS successfullyParsed is true

TEST COMPLETE
